添加上传下载功能

+++ network/src/main/java/tk/hongbo/network/helper/DownHelper.java
@@ -1,8 +1,5 @@
 package tk.hongbo.network.helper;
 
-import android.os.Environment;
-import android.support.annotation.NonNull;
-
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -13,6 +10,7 @@ import okhttp3.Callback;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
+import tk.hongbo.network.utils.LogHelper;
 
 public class DownHelper {
 
@@ -45,6 +43,7 @@ public class DownHelper {
             @Override
             public void onFailure(Call call, IOException e) {
                 // 下载失败
+                LogHelper.e("下载失败", e);
                 listener.onDownloadFailed();
             }
 
@@ -54,12 +53,15 @@ public class DownHelper {
                 byte[] buf = new byte[2048];
                 int len = 0;
                 FileOutputStream fos = null;
-                // 储存下载文件的目录
-                String savePath = isExistDir(new File(localUrl).getAbsoluteFile().getParent());
+                if (!isExistDir(localUrl)) {
+                    LogHelper.e("下载目录不存在", null);
+                    listener.onDownloadFailed();
+                    return;
+                }
                 try {
                     is = response.body().byteStream();
                     long total = response.body().contentLength();
-                    File file = new File(savePath, getNameFromUrl(url));
+                    File file = new File(localUrl);
                     fos = new FileOutputStream(file);
                     long sum = 0;
                     while ((len = is.read(buf)) != -1) {
@@ -71,8 +73,10 @@ public class DownHelper {
                     }
                     fos.flush();
                     // 下载完成
+                    LogHelper.d("下载成功");
                     listener.onDownloadSuccess();
                 } catch (Exception e) {
+                    LogHelper.e("下载写入文件异常", e);
                     listener.onDownloadFailed();
                 } finally {
                     try {
@@ -93,25 +97,14 @@ public class DownHelper {
     /**
      * @param saveDir
      * @return
-     * @throws IOException 判断下载目录是否存在
      */
-    private String isExistDir(String saveDir) throws IOException {
-        // 下载位置
-        File downloadFile = new File(Environment.getExternalStorageDirectory(), saveDir);
-        if (!downloadFile.mkdirs()) {
-            downloadFile.createNewFile();
+    private boolean isExistDir(String saveDir) {
+        File downloadFile = new File(saveDir);
+        File fileDir = new File(downloadFile.getAbsoluteFile().getParent());
+        if (!fileDir.exists()) {
+            return false;
         }
-        String savePath = downloadFile.getAbsolutePath();
-        return savePath;
-    }
-
-    /**
-     * @param url
-     * @return 从下载连接中解析出文件名
-     */
-    @NonNull
-    private String getNameFromUrl(String url) {
-        return url.substring(url.lastIndexOf("/") + 1);
+        return true;
     }
 
     public interface OnDownloadListener {
